+++ pajbot/modules/default_chat_states.py
@@ -95,13 +95,37 @@ class DefaultChatStatesModule(BaseModule):
             return True
 
         if self.settings["emoteonly"] == self.ONLINE_PHRASE:
-            self.bot.privmsg(".emoteonly")
+            try:
+                self.bot.twitch_helix_api.update_emote_only_mode(
+                    self.bot.streamer.id,
+                    self.bot.bot_user.id,
+                    self.bot.bot_token_manager,
+                    True,
+                )
+            except HTTPError as e:
+                if e.response.status_code == 401:
+                    log.error(f"Failed to update emote only mode, unauthorized: {e} - {e.response.text}")
+                    self.bot.send_message("Error: The bot must be re-authed in order to update emote only mode.")
+                else:
+                    log.error(f"Failed to update emote only mode: {e} - {e.response.text}")
 
         if self.settings["subonly"] == self.ONLINE_PHRASE:
             self.bot.privmsg(".subonly")
 
         if self.settings["r9k"] == self.ONLINE_PHRASE:
-            self.bot.privmsg(".uniquechat")
+            try:
+                self.bot.twitch_helix_api.update_unique_chat_mode(
+                    self.bot.streamer.id,
+                    self.bot.bot_user.id,
+                    self.bot.bot_token_manager,
+                    True,
+                )
+            except HTTPError as e:
+                if e.response.status_code == 401:
+                    log.error(f"Failed to update unique chat mode, unauthorized: {e} - {e.response.text}")
+                    self.bot.send_message("Error: The bot must be re-authed in order to update unique chat mode.")
+                else:
+                    log.error(f"Failed to update unique chat mode: {e} - {e.response.text}")
 
         if self.settings["slow_option"] == self.ONLINE_PHRASE:
             try:
@@ -133,13 +157,37 @@ class DefaultChatStatesModule(BaseModule):
             return True
 
         if self.settings["emoteonly"] == self.OFFLINE_PHRASE:
-            self.bot.privmsg(".emoteonly")
+            try:
+                self.bot.twitch_helix_api.update_emote_only_mode(
+                    self.bot.streamer.id,
+                    self.bot.bot_user.id,
+                    self.bot.bot_token_manager,
+                    True,
+                )
+            except HTTPError as e:
+                if e.response.status_code == 401:
+                    log.error(f"Failed to update emote only mode, unauthorized: {e} - {e.response.text}")
+                    self.bot.send_message("Error: The bot must be re-authed in order to update emote only mode.")
+                else:
+                    log.error(f"Failed to update emote only mode: {e} - {e.response.text}")
 
         if self.settings["subonly"] == self.OFFLINE_PHRASE:
             self.bot.privmsg(".subonly")
 
         if self.settings["r9k"] == self.OFFLINE_PHRASE:
-            self.bot.privmsg(".uniquechat")
+            try:
+                self.bot.twitch_helix_api.update_unique_chat_mode(
+                    self.bot.streamer.id,
+                    self.bot.bot_user.id,
+                    self.bot.bot_token_manager,
+                    True,
+                )
+            except HTTPError as e:
+                if e.response.status_code == 401:
+                    log.error(f"Failed to update unique chat mode, unauthorized: {e} - {e.response.text}")
+                    self.bot.send_message("Error: The bot must be re-authed in order to update unique chat mode.")
+                else:
+                    log.error(f"Failed to update unique chat mode: {e} - {e.response.text}")
 
         if self.settings["slow_option"] == self.OFFLINE_PHRASE:
             try: